[ownership] Extract out the main visitor SemanticARCOptVisitor from SemanticARCOpts.h into its own header.
This is so I can move individual sub-optimizations on SemanticARCOptVisitor into
their own files. So for instance, there will be one file containing the load
[copy] optimization and another containing the phi optimization, etc.
